Add skip modified after drush option.
Prevents the modification of a label or datastream if the datastream or object was modified
after this data. This should avoid losing data in the case the datastream or label was
changed unknowingly. The date should be chosen carefully, for example the date of the export
the csvfile was based on.